Portfolio/requirements-portfolio-pdf.txt
markdown>=3.4
Pillow>=9.0
playwright>=1.40
